+++ airbyte-server/src/main/java/io/airbyte/server/handlers/ConnectionsHandler.java
@@ -217,7 +217,9 @@ public class ConnectionsHandler {
     metadata.put("connector_destination_definition_id", destinationDefinition.getDestinationDefinitionId());
 
     final String frequencyString;
-    if (standardSync.getManual()) {
+    if (standardSync.getScheduleType() != null) {
+      frequencyString = getFrequencyStringFromScheduleType(standardSync.getScheduleType(), standardSync.getScheduleData());
+    } else if (standardSync.getManual()) {
       frequencyString = "manual";
     } else {
       final long intervalInMinutes = TimeUnit.SECONDS.toMinutes(ScheduleHelpers.getIntervalInSecond(standardSync.getSchedule()));
@@ -317,12 +319,8 @@ public class ConnectionsHandler {
     newStreams.forEach(((streamDescriptor, airbyteStreamConfiguration) -> {
       final AirbyteStreamConfiguration oldConfig = oldStreams.get(streamDescriptor);
 
-      if (oldConfig == null) {
-        // The stream is a new one, the config has not change and it needs to be in the schema change list.
-      } else {
-        if (haveConfigChange(oldConfig, airbyteStreamConfiguration)) {
-          streamWithDifferentConf.add(streamDescriptor);
-        }
+      if (oldConfig != null && haveConfigChange(oldConfig, airbyteStreamConfiguration)) {
+        streamWithDifferentConf.add(streamDescriptor);
       }
     }));
 
@@ -424,6 +422,24 @@ public class ConnectionsHandler {
       throws ConfigNotFoundException, IOException, JsonValidationException {
     final StandardSync standardSync = configRepository.getStandardSync(connectionId);
     return ApiPojoConverters.internalToConnectionRead(standardSync);
+  }
+
+  private static String getFrequencyStringFromScheduleType(final ScheduleType scheduleType, final ScheduleData scheduleData) {
+    switch (scheduleType) {
+      case MANUAL -> {
+        return "manual";
+      }
+      case BASIC_SCHEDULE -> {
+        return TimeUnit.SECONDS.toMinutes(ScheduleHelpers.getIntervalInSecond(scheduleData.getBasicSchedule())) + " min";
+      }
+      case CRON -> {
+        // TODO(https://github.com/airbytehq/airbyte/issues/2170): consider something more detailed.
+        return "cron";
+      }
+      default -> {
+        throw new RuntimeException("Unexpected schedule type");
+      }
+    }
   }
 
 }