+++ airbyte-tests/src/automaticMigrationAcceptanceTest/java/io/airbyte/test/automaticMigrationAcceptance/ImportApi.java
@@ -62,7 +62,7 @@ public class ImportApi {
       if (memberVarResponseInterceptor != null) {
         memberVarResponseInterceptor.accept(localVarResponse);
       }
-      if (localVarResponse.statusCode() / 100 != 2) {
+      if (errorResponse(localVarResponse)) {
         throw new ApiException(localVarResponse.statusCode(),
             "importArchive call received non-success response",
             localVarResponse.headers(),
@@ -79,6 +79,10 @@ public class ImportApi {
       Thread.currentThread().interrupt();
       throw new ApiException(e);
     }
+  }
+
+  private Boolean errorResponse(final HttpResponse<InputStream> localVarResponse) {
+    return localVarResponse.statusCode() / 100 != 2;
   }
 
   private HttpRequest.Builder importArchiveRequestBuilder(final File body) throws ApiException {