+++ airbyte-webapp/src/components/Table/Table.tsx
@@ -14,7 +14,7 @@ type IHeaderProps = {
   collapse?: boolean;
   customWidth?: number;
   customPadding?: PaddingProps;
-} & ColumnInstance;
+} & ColumnInstance<Record<string, unknown>>;
 
 type ICellProps = {
   column: IHeaderProps;
@@ -43,6 +43,9 @@ const Tr = styled.tr<{
 }>`
   background: ${({ theme, erroredRows }) => (erroredRows ? theme.dangerTransparentColor : theme.whiteColor)};
   cursor: ${({ hasClick }) => (hasClick ? "pointer" : "auto")};
+  &:hover {
+    background-color: ${({ theme }) => `${theme.grey50}`};
+  }
 `;
 
 const Td = styled.td<{
@@ -92,7 +95,7 @@ const Th = styled.th<IThProps>`
   }
 
   &:last-child {
-    padding-left: 45px;
+    padding-left: 15px;
     border-radius: 0 10px 0 0;
   }
 `;