+++ airbyte-webapp/src/components/base/Input/Input.test.tsx
@@ -1,8 +1,10 @@
-import { fireEvent } from "@testing-library/react";
-
-import { render } from "utils/testutils";
+import { fireEvent, waitFor } from "@testing-library/react";
+import { act } from "react-dom/test-utils";
+import { render } from "test-utils/testutils";
 
 import { Input } from "./Input";
+// eslint-disable-next-line css-modules/no-unused-class
+import styles from "./Input.module.scss";
 
 describe("<Input />", () => {
   test("renders text input", async () => {
@@ -39,9 +41,66 @@ describe("<Input />", () => {
 
     getByTestId("toggle-password-visibility-button")?.click();
 
-    expect(getByTestId("input")).toHaveAttribute("type", "text");
-    expect(getByTestId("input")).toHaveValue(value);
+    const inputEl = getByTestId("input") as HTMLInputElement;
+
+    expect(inputEl).toHaveAttribute("type", "text");
+    expect(inputEl).toHaveValue(value);
+    expect(inputEl.selectionStart).toBe(value.length);
     expect(getByRole("img", { hidden: true })).toHaveAttribute("data-icon", "eye-slash");
+  });
+
+  test("showing password should remember cursor position", async () => {
+    const value = "eight888";
+    const selectionStart = Math.round(value.length / 2);
+
+    const { getByTestId } = await render(<Input type="password" defaultValue={value} />);
+    const inputEl = getByTestId("input") as HTMLInputElement;
+
+    act(() => {
+      inputEl.selectionStart = selectionStart;
+    });
+
+    getByTestId("toggle-password-visibility-button")?.click();
+
+    expect(inputEl.selectionStart).toBe(selectionStart);
+  });
+
+  test("hides password on blur", async () => {
+    const value = "eight888";
+    const { getByTestId, getByRole } = await render(<Input type="password" defaultValue={value} />);
+
+    getByTestId("toggle-password-visibility-button").click();
+
+    const inputEl = getByTestId("input");
+
+    expect(inputEl).toHaveFocus();
+    act(() => inputEl.blur());
+
+    await waitFor(() => {
+      expect(inputEl).toHaveAttribute("type", "password");
+      expect(getByRole("img", { hidden: true })).toHaveAttribute("data-icon", "eye");
+    });
+  });
+
+  test("cursor position should be at the end after blur and and clicking on show password button", async () => {
+    const value = "eight888";
+    const { getByTestId } = await render(<Input type="password" defaultValue={value} />);
+    const inputEl = getByTestId("input") as HTMLInputElement;
+
+    getByTestId("toggle-password-visibility-button").click();
+    expect(inputEl).toHaveFocus();
+    act(() => {
+      inputEl.selectionStart = value.length / 2;
+      inputEl.blur();
+    });
+
+    await waitFor(() => {
+      expect(inputEl).toHaveAttribute("type", "password");
+    });
+
+    getByTestId("toggle-password-visibility-button").click();
+    expect(inputEl).toHaveFocus();
+    expect(inputEl.selectionStart).toBe(value.length);
   });
 
   test("should trigger onChange once", async () => {
@@ -60,7 +119,7 @@ describe("<Input />", () => {
     fireEvent.focus(inputEl);
     fireEvent.focus(inputEl);
 
-    expect(getByTestId("input-container")).toHaveClass("input-container--focused");
+    expect(getByTestId("input-container")).toHaveClass(styles.focused);
   });
 
   test("does not have focused class after blur", async () => {
@@ -71,7 +130,7 @@ describe("<Input />", () => {
     fireEvent.blur(inputEl);
     fireEvent.blur(inputEl);
 
-    expect(getByTestId("input-container")).not.toHaveClass("input-container--focused");
+    expect(getByTestId("input-container")).not.toHaveClass(styles.focused);
   });
 
   test("calls onFocus if passed as prop", async () => {