+++ airbyte-webapp/src/views/Connection/ConnectionForm/formConfig.tsx
@@ -26,10 +26,12 @@ import {
 } from "core/request/AirbyteClient";
 import { ValuesProps } from "hooks/services/useConnectionHook";
 import { useCurrentWorkspace } from "services/workspaces/WorkspacesService";
+import { isValidCron } from "utils/cron";
 
 import calculateInitialCatalog from "./calculateInitialCatalog";
+import { ConnectionOrPartialConnection } from "./ConnectionForm";
 
-interface FormikConnectionFormValues {
+export interface FormikConnectionFormValues {
   name?: string;
   scheduleType?: ConnectionScheduleType | null;
   scheduleData?: ConnectionScheduleData | null;
@@ -41,9 +43,9 @@ interface FormikConnectionFormValues {
   normalization?: NormalizationType;
 }
 
-type ConnectionFormValues = ValuesProps;
+export type ConnectionFormValues = ValuesProps;
 
-const SUPPORTED_MODES: Array<[SyncMode, DestinationSyncMode]> = [
+export const SUPPORTED_MODES: Array<[SyncMode, DestinationSyncMode]> = [
   [SyncMode.incremental, DestinationSyncMode.append_dedup],
   [SyncMode.full_refresh, DestinationSyncMode.overwrite],
   [SyncMode.incremental, DestinationSyncMode.append],
@@ -57,7 +59,7 @@ const DEFAULT_SCHEDULE: ConnectionScheduleData = {
   },
 };
 
-function useDefaultTransformation(): OperationCreate {
+export function useDefaultTransformation(): OperationCreate {
   const workspace = useCurrentWorkspace();
   return {
     name: "My dbt transformations",
@@ -73,22 +75,51 @@ function useDefaultTransformation(): OperationCreate {
   };
 }
 
-const connectionValidationSchema = yup
+export const connectionValidationSchema = yup
   .object({
     name: yup.string().required("form.empty.error"),
-    scheduleType: yup.string().oneOf([ConnectionScheduleType.manual, ConnectionScheduleType.basic]),
-    scheduleData: yup
-      .object({
-        basicSchedule: yup
+    scheduleType: yup
+      .string()
+      .oneOf([ConnectionScheduleType.manual, ConnectionScheduleType.basic, ConnectionScheduleType.cron]),
+    scheduleData: yup.mixed().when("scheduleType", (scheduleType) => {
+      if (scheduleType === ConnectionScheduleType.basic) {
+        return yup.object({
+          basicSchedule: yup
+            .object({
+              units: yup.number().required("form.empty.error"),
+              timeUnit: yup.string().required("form.empty.error"),
+            })
+            .defined("form.empty.error"),
+        });
+      } else if (scheduleType === ConnectionScheduleType.manual) {
+        return yup.mixed().notRequired();
+      }
+
+      return yup.object({
+        cron: yup
           .object({
-            units: yup.number().required("form.empty.error"),
-            timeUnit: yup.string().required("form.empty.error"),
+            cronExpression: yup
+              .string()
+              .test({
+                name: "cronExpression",
+                message: "Incorrect cron string",
+                test(value) {
+                  if (value && isValidCron(value, { seconds: true, allowBlankDay: true })) {
+                    return true;
+                  }
+
+                  return this.createError({
+                    message: "form.cron.invalid",
+                    path: `cronExpression`,
+                  });
+                },
+              })
+              .required("form.empty.error"),
+            cronTimeZone: yup.string().required("form.empty.error"),
           })
-          .nullable()
           .defined("form.empty.error"),
-      })
-      .nullable()
-      .defined("form.empty.error"),
+      });
+    }),
     namespaceDefinition: yup
       .string()
       .oneOf([
@@ -172,7 +203,7 @@ const connectionValidationSchema = yup
  * @param initialOperations
  * @param workspaceId
  */
-function mapFormPropsToOperation(
+export function mapFormPropsToOperation(
   values: {
     transformations?: OperationRead[];
     normalization?: NormalizationType;
@@ -210,10 +241,10 @@ function mapFormPropsToOperation(
   return newOperations;
 }
 
-const getInitialTransformations = (operations: OperationCreate[]): OperationRead[] =>
+export const getInitialTransformations = (operations: OperationCreate[]): OperationRead[] =>
   operations?.filter(isDbtTransformation) ?? [];
 
-const getInitialNormalization = (
+export const getInitialNormalization = (
   operations?: Array<OperationRead | OperationCreate>,
   isEditMode?: boolean
 ): NormalizationType => {
@@ -227,10 +258,8 @@ const getInitialNormalization = (
     : NormalizationType.basic;
 };
 
-const useInitialValues = (
-  connection:
-    | WebBackendConnectionRead
-    | (Partial<WebBackendConnectionRead> & Pick<WebBackendConnectionRead, "syncCatalog" | "source" | "destination">),
+export const useInitialValues = (
+  connection: ConnectionOrPartialConnection,
   destDefinition: DestinationDefinitionSpecificationRead,
   isEditMode?: boolean
 ): FormikConnectionFormValues => {
@@ -244,6 +273,7 @@ const useInitialValues = (
     const initialValues: FormikConnectionFormValues = {
       name: connection.name ?? `${connection.source.name} <> ${connection.destination.name}`,
       syncCatalog: initialSchema,
+      scheduleType: connection.connectionId ? connection.scheduleType : ConnectionScheduleType.basic,
       scheduleData: connection.connectionId ? connection.scheduleData ?? null : DEFAULT_SCHEDULE,
       prefix: connection.prefix || "",
       namespaceDefinition: connection.namespaceDefinition || NamespaceDefinitionType.source,
@@ -261,10 +291,25 @@ const useInitialValues = (
     }
 
     return initialValues;
-  }, [initialSchema, connection, isEditMode, destDefinition]);
+  }, [
+    connection.connectionId,
+    connection.destination.name,
+    connection.name,
+    connection.namespaceDefinition,
+    connection.namespaceFormat,
+    connection.operations,
+    connection.prefix,
+    connection.scheduleData,
+    connection.scheduleType,
+    connection.source.name,
+    destDefinition.supportsDbt,
+    destDefinition.supportsNormalization,
+    initialSchema,
+    isEditMode,
+  ]);
 };
 
-const useFrequencyDropdownData = (
+export const useFrequencyDropdownData = (
   additionalFrequency: WebBackendConnectionRead["scheduleData"]
 ): DropDownRow.IDataItem[] => {
   const { formatMessage } = useIntl();
@@ -282,28 +327,34 @@ const useFrequencyDropdownData = (
       }
     }
 
-    return frequencies.map((frequency) => ({
+    const basicFrequencies = frequencies.map((frequency) => ({
       value: frequency,
-      label: frequency
-        ? formatMessage(
-            {
-              id: `form.every.${frequency.timeUnit}`,
-            },
-            { value: frequency.units }
-          )
-        : formatMessage({ id: "frequency.manual" }),
+      label: formatMessage(
+        {
+          id: `form.every.${frequency.timeUnit}`,
+        },
+        { value: frequency.units }
+      ),
     }));
-  }, [formatMessage, additionalFrequency]);
-};
 
-export type { ConnectionFormValues, FormikConnectionFormValues };
-export {
-  connectionValidationSchema,
-  useInitialValues,
-  useFrequencyDropdownData,
-  mapFormPropsToOperation,
-  SUPPORTED_MODES,
-  useDefaultTransformation,
-  getInitialNormalization,
-  getInitialTransformations,
+    // Add Manual and Custom to the frequencies list
+    const customFrequency = formatMessage({
+      id: "frequency.cron",
+    });
+    const manualFrequency = formatMessage({
+      id: "frequency.manual",
+    });
+    const otherFrequencies = [
+      {
+        label: manualFrequency,
+        value: manualFrequency.toLowerCase(),
+      },
+      {
+        label: customFrequency,
+        value: customFrequency.toLowerCase(),
+      },
+    ];
+
+    return [...otherFrequencies, ...basicFrequencies];
+  }, [formatMessage, additionalFrequency]);
 };