+++ airbyte-workers/src/main/java/io/airbyte/workers/temporal/scheduling/activities/JobCreationAndStatusUpdateActivityImpl.java
@@ -13,9 +13,11 @@ import io.airbyte.config.DestinationConnection;
 import io.airbyte.config.FailureReason;
 import io.airbyte.config.JobOutput;
 import io.airbyte.config.JobSyncConfig;
+import io.airbyte.config.NormalizationSummary;
 import io.airbyte.config.StandardDestinationDefinition;
 import io.airbyte.config.StandardSync;
 import io.airbyte.config.StandardSyncOperation;
+import io.airbyte.config.SyncStats;
 import io.airbyte.config.helpers.LogClientSingleton;
 import io.airbyte.config.helpers.LogConfigs;
 import io.airbyte.config.persistence.ConfigNotFoundException;
@@ -43,29 +45,48 @@ import io.airbyte.workers.helper.FailureHelper;
 import io.airbyte.workers.run.TemporalWorkerRunFactory;
 import io.airbyte.workers.run.WorkerRun;
 import io.airbyte.workers.temporal.exception.RetryableException;
+import io.micronaut.context.annotation.Requires;
 import java.io.IOException;
 import java.nio.file.Path;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
+import javax.inject.Inject;
+import javax.inject.Singleton;
 import lombok.AllArgsConstructor;
+import lombok.NoArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
 @AllArgsConstructor
+@NoArgsConstructor
 @Slf4j
+@Singleton
+@Requires(property = "airbyte.worker.plane",
+          notEquals = "DATA_PLANE")
 public class JobCreationAndStatusUpdateActivityImpl implements JobCreationAndStatusUpdateActivity {
 
-  private final SyncJobFactory jobFactory;
-  private final JobPersistence jobPersistence;
-  private final TemporalWorkerRunFactory temporalWorkerRunFactory;
-  private final WorkerEnvironment workerEnvironment;
-  private final LogConfigs logConfigs;
-  private final JobNotifier jobNotifier;
-  private final JobTracker jobTracker;
-  private final ConfigRepository configRepository;
-  private final JobCreator jobCreator;
-  private final StreamResetPersistence streamResetPersistence;
-  private final JobErrorReporter jobErrorReporter;
+  @Inject
+  private SyncJobFactory jobFactory;
+  @Inject
+  private JobPersistence jobPersistence;
+  @Inject
+  private TemporalWorkerRunFactory temporalWorkerRunFactory;
+  @Inject
+  private WorkerEnvironment workerEnvironment;
+  @Inject
+  private LogConfigs logConfigs;
+  @Inject
+  private JobNotifier jobNotifier;
+  @Inject
+  private JobTracker jobTracker;
+  @Inject
+  private ConfigRepository configRepository;
+  @Inject
+  private JobCreator jobCreator;
+  @Inject
+  private StreamResetPersistence streamResetPersistence;
+  @Inject
+  private JobErrorReporter jobErrorReporter;
 
   @Override
   public JobCreationOutput createNewJob(final JobCreationInput input) {
@@ -172,7 +193,9 @@ public class JobCreationAndStatusUpdateActivityImpl implements JobCreationAndSta
 
       if (input.getStandardSyncOutput() != null) {
         final JobOutput jobOutput = new JobOutput().withSync(input.getStandardSyncOutput());
-        jobPersistence.writeOutput(jobId, attemptId, jobOutput);
+        final SyncStats syncStats = jobOutput.getSync().getStandardSyncSummary().getTotalStats();
+        final NormalizationSummary normalizationSummary = jobOutput.getSync().getNormalizationSummary();
+        jobPersistence.writeOutput(jobId, attemptId, jobOutput, syncStats, normalizationSummary);
       } else {
         log.warn("The job {} doesn't have any output for the attempt {}", jobId, attemptId);
       }
@@ -230,7 +253,9 @@ public class JobCreationAndStatusUpdateActivityImpl implements JobCreationAndSta
 
       if (input.getStandardSyncOutput() != null) {
         final JobOutput jobOutput = new JobOutput().withSync(input.getStandardSyncOutput());
-        jobPersistence.writeOutput(jobId, attemptId, jobOutput);
+        final SyncStats syncStats = jobOutput.getSync().getStandardSyncSummary().getTotalStats();
+        final NormalizationSummary normalizationSummary = jobOutput.getSync().getNormalizationSummary();
+        jobPersistence.writeOutput(jobId, attemptId, jobOutput, syncStats, normalizationSummary);
       }
 
       emitJobIdToReleaseStagesMetric(OssMetricsRegistry.ATTEMPT_FAILED_BY_RELEASE_STAGE, jobId);